// src/types/navigation.types.ts
export interface CategoryItems {
  NEWS: string;
  REVIEWS: string;
  SPORT_CALENDAR: string;
  CONFERENCES: string;
  SERVICES: string;
  PROMOCODES: string;
  KNOWLEDGE_BASE: string;
  IGAMING_JOBS: string;
  IGAMING_COMPANIES: string;
}

export interface ConferenceItems {
  ALL: string;
  ADD: string;
}

export interface ServiceItems {
  BLOCKED: string;
  SPORTS_WIDGET: string;
  EVENTS_WIDGET: string;
}

export interface PromocodeItems {
  OPERATORS: string;
  AFFILIATES: string;
  IGAMING_COMPANIES: string;
  ADD: string;
}

export interface JobItems {
  ALL: string;
  ADD: string;
}

export interface CompanyItems {
  ALL: string;
  ADD: string;
}